Reproducible Python environment for StarkNet contracts
Had a bit of trouble while setting up the environment to run the StarkNet tests, so added the Poetry dependency manager to make the Python env more deterministic and reproducible. I've tried out several Python dep managers over the years and have used Pipenv and Poetry seriously. Of those two I have a mild preference for Poetry (faster, and also feels like a higher-quality project in general). But fine with using another dependency manager if you prefer.

I also configured the project to use a local venv instead of the global one in ~/cairo_venv. You mentioned that would be nice and I agree it's a better pattern. Seems like the tooling supports it, I just had to change a couple of config options in the StarkWare VS Code extension to get it working.