Serialize arbitrarily-sized byte arrays to Felts
Since Starknet only gives us Felts to work with for on-chain storage, I wrote a somewhat homebrew serialization library that takes an arbitrary array of bytes and serializes/deserializes it to Felts.

I think it would be helpful to get another pair of eyes on this -- the code ended up both longer and more complex than I expected, so would appreciate a sanity check that I'm not missing an easier way to do this and didn't included any obvious bugs.