Refactor Redwood<->Starknet interface
So this was instigated as part of a larger change to start verifying notaries using their StarkNet wallets. However, we've decided to move away from that design so I stripped out the parts of this related to that. I still think though that the `starknet.ts` file was getting a bit unwieldy, so going to commit this diff that refactors it a bit.