Use sign-in with Ethereum and Redwood auth
Allow users to authenticate themselves by signing a server-provided one-time-use string.

This is useful for allowing notaries to submit an "intention to notarize" client-side that we can then trust to perform the actual notarization server-side.

Our original plan was to perform all notarization and profile submission client-side, but unfortunately StarkNet is being slow and unreliable and so we need to be able to track and retry failed submissions, which this allows us to do more easily.