Fix `useLazyQuery` error
I'm not 100% sure why `useLazyQuery` was throwing an error -- after checking the docs and other places I've used it I'm pretty sure it shouldn't have failed. But I couldn't get to the bottom of the issue, so I worked around it by switching to a `useQuery`+`skip` call that should have the same behavior.