Fix bug in `useGuard` with dynamic routes
Turns out that the key used in the RW router loading state doesn't interpolate variables. I think this might lead to bugs in the RW router if you're navigating between two routes that are identical except for the interpolated variables, but I'm not going to worry about their potential bugs and will just fix my own.