Dummy submission signature
This adds a dummy signature step to the profile submission so we can see how users react, whether the language is confusing, etc.

Unfortunately the UX of submission signing will probably need to change before this goes live for real. I'm not sure whether it's possible to generate short IPFS hashes locally (the `ipfs-only-hash` project, which ought to handle it, doesn't seem to deal with the `mhlen` argument we need to generate custom short hashes that'll fit in a felt). If it isn't possible, we may need to rework the UX so we can generate the string to sign *after* the photo and video have been uploaded.

Another potential tech-driven UX constraint is that I'm not sure how annoying it'll be to verify the signature of a long string in Cairo. If that ends up being hard to pull off, we may have to settle for just signing the bare profile CID, which means we'll have to find a way to indicate to the user what the heck they're signing.