Use derived wallets and authentication on all pages
Big change to our frontend model -- you now always have to sign a string before you can sign in. We use the signature to create an in-browser wallet. This disconnects your Zorro identity from the wallet you connected with. The downside is that it also creates a credential that can be more easily stolen (the signature). But feels like the right tradeoff given the options.